[FIX] account_reports: traceback during print of financial report
This restore the expression from saas-11 that was changed in rev 0c8fc3417547d93a0a5ed1f52ef03318e5c9acc6
that was doing two things: removing the groups defined on template tags
and -what concerns us- replacing empty string in t-att-* attributes by
None. This was done because the new QWeb will completely remove the tags
if they are None and the resulting DOM is cleaner.

However in this special case it remplaces an empty string inside a join:
 "' '.join(['a', 'b', None])" raises a TypeError exception and it is our
traceback.