[FIX] account_reports and l10n_.._reports : adapt the tax reports to commit 39d68ac (in community repo) to display correctly the journal items involved in operation using cash basis taxes.
Previsouly they were directly included in the tax reports, in the period of the invoice (wrong), and even if it was still unpaid (wrong again).

Although this is a bugfix, it involves creation of new fields and thus cannot lands at it on stable versions where our policy disallow database changes.