[FIX] delivery_ups: label rotation
When a the UPS label is generated, ti is automatically rotated and
converted to PDF. However, the rotation keeps by default the same aspect
ratio, meaning that the label will be cropped

By using the option `expand`, the aspect ratio will be changed. In
practic, it converts a 1400x800 pixels image to a 800x1400 pixels image.

opw-686286