[REF] sale_contract: split model (subscription <-> template)
Using the same technical objects for subscriptions and subscription templates
has several disadvantages for templates:
    * company field is mandatory (which is a bummer, because intercompany
      ir.rules for analytic accounts apply)
    * pricelist field does not mean anything in and of itself (especially
      in multicompany/multicurrency setups)
    * xml code for views is complicated (attrs everywhere)
    * some fields conflict between each other (state vs. active)

For this reason, we decided to split the current model into the template
model and the subscription model.

The subscription model remains mostly unchanged (some template-specific
fields are removed and the behaviour is adapted to the new template model).

Some new models with a very limited amount of logic are created (ACLs and
ir_rules included).