[IMP] *_contract: use related fields for recurrence and remove manager_id field
This ensures that there will not be any loss of synchronization
between the recurring period of a subscription and that of its
template.

The field manager_id field is kind of a duplicate of user_id and
is thus removed (can somebody tell the difference between a
Responsible and a Sales Representative?).