[FIX] web_enterprise: scroll in settings view
The issue is that when a settings view is big enough to be scrolled, the
background color changes from white to grey as soon as it exceeds the
screen size. This is visible in the inventory app settings. This is due
to the fact that the class `o_form_nosheet` is applied and forces an
height:100%.

This is related to rev 4133626d5c9244a2db0f7d2bddf2de34af6e8745
Before this commit, the `o_form_nosheet` has an height:100% and
an overflow:auto. This commit removes the overflow:auto and thus
the height:100% rule becomes problematic as text can now outgrows
the container.

As it is the action manager that scrolls, we should let the
elements inside the view manager grow as they should. On top of
that, a min-height rule is already set to fill at lest the screen
if the content is not big enough.