[FIX] delivery_ups: do not rotate label
The previous UPS commit fixes the rotation of the image. However, having
a PDF file in portrait will be printed on full page on most system,
making the label way too big.

The fix is to not rotate the label, so it's easier for the user to print
it with the correct size.

opw-686286