[FIX] account_batch_deposit: printing report
When we are trying to print a batch deposit, we get a traceback instead
of a pdf. This is due to the new way to call render_html in get_html
introduced in rev https://github.com/odoo/odoo/commit/a3f83916d but
was not applied for this report.

To fix this, we changed the 'render_html' method decorators and signature
and used the 'docids' parameter to generate reports.