[FIX] account_reports: `date_from` in context
Go to Accounting >> Reports >> Executive Summary >> click on 'CASH'. A
traceback appears. There is no error when accessing directly to the
report thanks to Accounting >> Reports >> Cash Flow Statement.

This is because the 'CASHSTART' line requires the beginning of the
period to be defined (`date_from`), and in this case, it is not.

opw-687558