[FIX] website_sign: translation of emails
When sending a document thanks to e-Sign, the e-mail is always in
English, while it should be in the language of the recipient. Moreover,
the link in the e-mail always redirect to the English website.

To solve this, we add the lang of the partner in the context when
browsing the template, so it is translated.

Regarding the English website, this is more tricky to introduce the
appropriate lang in the URL without risking to break the link (website
not translated, multi-website...). Therefore, we just add the language
selector in the appropriate template of the frontend.

opw-685806