[FIX] account_reports, cash flow report: display lines not in a journal of type bank/cash at the date of the payment instead of at the date of the invoice.
We use a magic trick to define a WITH statement as prologue to the usual queries returned by query_get() for the cashflow report, in order to shadow the table account_move_line entirely and let the query_get() work normally although we're fetching rows from the temporary table (built in the WITH statement) instead of the account_move_line table.

NOTE: This is totally unusual for implementing a complex report but has been done as a bugfix when we realized the generic report was showing the invoice's account under the _invoice_ date in the reporting (instead of payment date), and it was too late to implement that with a dedicated custom report.