[IMP] quality: no quick add in the quality alert calendar
Before this patch, when you wanted to create a quality alert from
the calendar view, you were presented with a wizard with only the
name field to fill. As this was not enough to create a quality
alert, the full form view was openned afterwards. This patch
opens the full form view directly.