[IMP] quality: quality check wizard and measures
There was a huge usability issue as the quality check wizard contains only
one field and, in the case of measure, the buttons in the footer were
updated with an onchange in order to display a `fail` or `pass` button, which
triggers a different method. So you **had** to click anywhere on the wizard
after filling the measure in order to display the right button.

This fix workaround this issue by displaying a single button `record` that
will call either `do_pass` or `do_fail` according to the recorded measure