[IMP] quality: don't bypass quality checks when validating a draft picking
If you have a draft picking and you validate it directly, the quality
check mechanism is totally bypassed. It's because the method that was
creating the quality checks was called in an override of "do_new_transfer"
(`validate` button) before actually calling super, so there was nothing
to do. But if you click on the "action_confirm" (`make as to do`) button
then the quality checks were created so it'll correctly work when you
finally click on validate.

To fix this issue we move the quality checks presence to to a lower
level, in the action_done of the stock move, so it is called during the
validate process. If it fails, the valide process is aborted and a
message is displayed to the user telling him he needs to do the
quality checks.