[FIX] account_bank_statement_import_csv: fix import
The method overriden `_parse_import_data` did call `super`,
but forget to return the result of the super.

Because of it, it was no longer possible to import anything else
than bank statement as soon as this module was installed.

e.g. it was no longer possible to import leads.

opw-688306