[FIX] account_yodlee: fix discontinuity in statement + some other errors
Previously we were using the balance yodlee was sending as ending balance for the bank statement and we were computing the starting balance doing the following: end balance minus sum of all transactions fetched by yodlee. However due to the fact that ending balance sent by yodlee sometimes include pending transaction and we only fetch posted transaction, we have discontinuity in the bank statement in Odoo. To prevent this problem, we changed the way to compute starting and ending balance doing the following: starting balance is the value of previous statement ending balance and ending balance is: starting balance + sum of all transactions fetched by yodlee.

Other fixes:
in case site is not supported by yodlee anymore, throw a nice error instead of a tarceback
Sometimes transactions does not have any communication, add a default value for those to prevent a traceback.