[FIX] crm_voip: don't override FieldChar events
FieldPhone is extending FieldEmail which is extending FieldChar.

So when we initialized a phone field adding a click handler, this
handler would get used for every field inheriting from FieldChar
(which had not cloned the events object).

For example in the contact view form, link and email fields would not be
clickable.

opw-686430