[IMP] website_helpdesk: notification email update
This commit is linked to a community commit. This commit remove the sudo used
in the mail/view controller. get_access_action must now correctly handle the
user to return the redirection action.

We also improve notification email recipients grouping and effectively add an
access button to tickets for customers in notification emails.