[FIX] account_reports: multi-company in reports
The folowing configuration leads to an issue:
- Set the instance to multicompany through Admin rights
- Create two users (User A and User B)
- Create two companies (company A and company B)
- Set company A currency on "EUR"
- Set company B currency on "AUD"
- Set company B as the child of company A
- Install chart of accounts for company B
- Set a currency rate under company A (1.5 on AUD)
- User B is under "current company B" but has in allowed company :
  "Company A" & "Company B"
- User A is under "current company A" but has in allowed company :
  "Company A" & "Company B"
- User B creates a customer invoice
- Check Chart of Account of company A
- Check Chart of Account of company B
- Check General Ledger of company A
- Check General Ledger of company B

What is the current behavior?
- The Chart of Account isn't applying the rate correctly
- Amounts shown on both Chart of Account are similar but in different
  currencies

The following reports do not take a currency table: General Ledger,
Trial Balance (Chart of Account) and the generic Tax Report. Ideally, we
should introduce the currency table in these reports. However, there is
a high risk to introduce side effects, while the added value would be
limited. Indeed, in these reports, we do not aggregate the results
between companies, since the chart of account is specific to a given
company.

Therefore, we prevent the user to select companies which have a currency
which is different from the currency of their company.

opw-687512