[IMP] web_enterprise: hide empty mobile action menu
The action menu in mobile contains the action buttons
(e.g. "create invoice" on sale order) but in some instances some action
buttons can be invisible depending on a given set of circumstances.

It is non-trivial to hide the action menu if these invisible buttons are
present: hence the action menu was always displayed despite being
possibly empty.

This commit hides the action menu in the particular situation that there
is zero action button in the view (the menu will still be displayed if
there is one invisible action button).

opw-688677