[FIX] website_helpdesk: remove and adapt dead code
The website_helpdesk module was adding a "New Team" button in the
frontend but it was not doing anything. This commit removes the
associated xml and JS code.
Also, the JS code contained initialization code which should have
been there for every website visitor. This was adapted and placed
in the website frontend assets instead of editor assets.