[FIX] mrp_plm, quality: plm and quality access rights
This fix puts the access rights for the plm/quality user/manager.
The stock user is automatically quality user as is the mrp user.
We need also ir.rules for company access rights.
The quality user should at least have access to the basic objects
for the case he is not stock user, but just belongs to the quality department.