[FXI] mail_gcm: import gcm external dependence
When we are trying to create a database in enterprise without having
the library "gcm" installed on the server, the database initiallisation
crashed.

The server meets errors, because odoo load all modules, even those who
are not installed and create the static routes.

To fix this bug, we made a 'try' of the import of gcm in the addon, has
it's done in 'base_import' or 'res_partner' for example.