[ADD] web_studio: new view renderers
This commit introduces the basic infrastructure to separate the
rendering from the views. This will be the base for the future
implementation of views. This contains form and list renderers.

Note that this new design uses 'model' to communicate with the
server.

Joint work with aab, adr, dmo, ged and mge.