[ADD] web_studio: kanban renderer and relational fields
This commit introduces the Kanban renderer and the implementation
of relational field widgets. Note that the Kanban renderer doesn't
uses those widgets yet, but it will someday.

Old view renderers are used by studio to offer a preview of views
for which the rendering isn't separated from the view yet.

Joint work with aab, adr, dmo, ged and mge.