[FIX] web_studio: form_editor: add a chatter
Activate mail_thread inheritance for custom models for which we want
to add a chatter on a form view, by setting to true the recently
added flag 'mail_thread' (84d7c6325).

Also ensure that the 'Add a Chatter' hook is only appended to the
form editor if the form view contains a sheet.