[FIX] web_studio: editors: don't save reverted failed operation
Operations can be undone/redone. When the user does an operation that
breaks the field_views inheritance, this operation is automatically
undone.

Before this fix, it was saved anyway in the stack of undone operations,
meaning that the user can press the 'Redo' button to redo it, but
obviously with the same result. So this fix simply forget this
reverted, wrong operation.

Also don't pass 'this' as argument when triggering the redo/undo related
events on the bus, as it isn't used by the handlers. This prevents
keeping somewhere a reference to a destroyed widget which can't thus
be garbage collected.