[FIX] web_studio: Correctly compute xpath when no attrs
In some case, we don't have any robust attrs (like:
name, for,...) to create an accurate xpath.

The first idea was to use 'index':
For example:
<group>
	<group string="child_1"/>
	<group string="child_2"/>
</group>
//group[1] will take the parent
//group[2] will take the second child
But, with the previous algorithm, it was not possible to have the first child.
In fact, '//group[1]' takes the parent and the first child but the framework
take only the first result.

Now, we add full xpath to a node in case of no robust attrs.