[FIX] web_studio: xml_editor: don't save reverted changes in views
Before this fix, when the user edited a view with the xml_editor and
then clicked on 'Save', the editor notified the change by triggering
up the former arch and the new arch, and updated its local reference
to the edited view, without taking into account the fact that the
change could broke the views inheritance, and thus be directly
reverted. In the case of a revert, the opposite operation is done,
and the former arch given by the xml_editor is restored.

This means that in this situation, if the user clicked a second time
on 'Save', the xml_editor notified the change with the former arch
being now the broken one, meaning that the correct arch couldn't be
restored anymore.

This rev. ensures that the xml_editor updates its local reference
to the arch only in case of a successful operation.