[FIX] web_enterprise: don't update menu_id with submenu's ids
The menu_id in the url represents the current primary menu id (i.e.
the id of the current app). When clicking on a submenu, we thus don't
want to update the menu_id in the url with that submenu's id, because
if we do, we will lose the current app information when reloading the
page, and thus all the menus in the navbar.

This was a problem when we selected a submenu directly from the
app switcher (using the search bar), as in that particular case,
the menu_id was set in the event triggered up (which wasn't the
case if the user simply clicked in a submenu in the navbar).