[FIX] web_studio: prevent unnecessary calls to /web_studio/init
An initial RPC is done when entering studio to load some required
information. This RPC was performed in the willStart of the studio
client action, which can be re-executed while already being in
studio (e.g. when clicking on 'Views' in the second navbar).

This commit moves that RPC in the function called when entering
studio.