[IMP] mail_fcm: compliant for saas module
In this commit we are defining a private method to get the fcm
credentials, this method is needed to have the availability to
override it in an internal module to provide the odoo.com credentials.
This method is private to avoid any rpc call on the credentials.

We also give a name to the res_config group in order to remove it on
the saas, where the odoo.com key will be used.