[FIX] web_studio: prevent crash when opening custom apps
The custom modules created by odoo studio were not created with a valid
icon.  This commit adds the icon, and some other metadata, which
prevents a crash when opening the record in the kanban view of
applications.

Sadly, for now, it only uses the default icon.  It would be better if we
could use the same icon as the ir ui menu of the application.