[FIX] web_studio: relational fields referencing parent in context
Temporary don't use field context when loading data to prevent a traceback
on the invoice form view (when there is a tax set in an invoice line).

Problem is due to the context set on the invoice_line_tax_ids field
("{'type': parent.type}"), as it makes reference to the parent, which
is not supported yet.