[FIX] web_studio: solves focus problems in form editor
When editing nodes in the form editor, the form renderer is
reinstantiated, loses all local state, click on the active pages, which
means that at the end, the view editor consider that the user is editing
a notebook page.

This commit changes this logic so that the clicks on pages are ignored
(from the point of view of the renderer, bootstrap still does its
magic).  This is not the perfect solution, since the current node being
edited will lose its focus in the editor, but still be edited in the
sidebar.