[FIX] helpdesk: use of invalid registry after install
After the install of modules done through
the method `_check_modules_to_install`,
the environment/registry with which the install was done
initially is no longer up-to-date,
and it should therefore no longer be used,
as it could lead to the use of an object instance
which is no longer valid.

This solves an issue regarding the activation
of all helpdesk team features, for which the field
`reS.users.share` was mentioned as no longer
in the computed fields of the model `res.users`:
it was, but no longer with the same object instance,
with the same memory address.

```
File "/home/odoo/odoo_dev/odoo/odoo/models.py", line 5267, in recompute
  fs = self.env[field.model_name]._field_computed[field]
  KeyError: res.users.share
```