[FIX] web_studio: model: correctly fetch selection_info in list
The postprocessing of fields (in charge of loading selection, status,
many2many... information, if requested by the fields_view) was only
done for form views. For instance, if there was a many2one field with
widget="selection" in a list view, there was a crash because the
selection_info (what to display) wasn't loaded.