[FIX] web_studio: avoid to generate button name each time
Before this commit, the name was generated in python after each
modification in form view.
The problem is that the button was recreated with a different name
but the operations list was always based on the old name.

Now, we avoid to change this name. The name is generated in js and
saved directly in the operations list.