[FIX] web_studio: OldViewRenderers: inherit from AbstractRenderer
The OldViewRenderers, providing an inert rendering of views for which the
rendering part hasn't been extracted yet (e.g. pivot, graph, calendar),
is manipulated by the ViewEditorManager like a classical Renderer. For
that reason, it should have the same API, and thus inherit from
AbstractRenderer.

This prevents a traceback 'Uncaught TypeError: this.editor.get_local_state
is not a function' when opening the xml editor.