[FIX] stock_barcode, mrp_barcode: barcode save
Since the add of mrp_barcode, a barcode to save the form has been added
but it doesn't work. This barcode doesn't work, because it represent
the string 'O-BTN.save', and the code handles the string 'O-CMD.SAVE'.

So to fix this, we have regenerated the barcodes, with the good string
'O-CMD-SAVE'.

This bug has been introduced in rev: https://github.com/odoo/enterprise/commit/0ba37243984e4c31416e096abe6a17b1d23117e7