[FIX] web_studio: correctly put studio info in context on refresh
We write in the user_context that we are in Studio so that the
ir.model.data of the customized resources are flagged with Studio.

However, this wasn't done at the right place, as the context wasn't
correctly updated after a refresh in studio (toggle_studio_mode() is
only called when the user clicks on the studio icon in the navbar).