[FIX] UnicodeWarning on account_reports install
Fix non-ASCII content in a Field attribute, which would be decoded
before storing it into the db (ir.model.fields) but not when comparing
the original value and the current one to know if the in-db value should
be updated, leading to a UnicodeWarning and spurious DB updates.

Could be fixed by either making the value a unicode literal, or by
making the string pure ASCII. Considering the content in question (weird
quotes), converting to ASCII seems like the most sensible change.