[FIX] mail_push: on delete cascade on bad model
The "ondelete=cascade" set on the field 'device_identity_ids' in
the 'res.partner' model doesn't do anything because it's not implemented
on One2many fields.

To perform expected behavior, we set "ondelete=cascade" on the field
'partner_id' in 'mail_push.device' model to delete the devices of a
partner if this one is deleted.