[FIX] stock_barcode: PickingBarcodeHandler has no return statement
The function 'stock_barcode.PickingBarcodeHandler' has no return statement
so when we are calling it trough a require call to get a variable to work with,
it doesn't return anything and it's so impossible to extend it in another
module.

To fix this problem, we added a return statement which returns the
'PickingBarcodeHandler' variable.