[FIX] web_studio: correctly style all dialogs in studio
Former solution required to manually add 'o_web_studio' classname on
every created dialog. It means that some dialogs weren't correctly
flagged, and that some of them *can't* be flagged easily (e.g. actions
in target new).

This new solution simply adds the classname 'o_in_studio' on the webclient
when we are in studio, and all dialogs are styled accordingly when that
classname is there.

Same mechanism is used to style the app switcher and the menu as well.

Finally, rename the 'o_web_studio' classname of the studio client action
into 'o_web_studio_client_action' to prevent unfortunate collisions.