[FIX] web_studio: flag all customizations done in Studio
The ir.model.data of every customization done in studio is flagged
(studio = True), so that we can retrieve all customizations and
export them. This is done by overriding the write() and create()
methods of the models we can customize. However, this wasn't done
yet for several models.

Instead of manually overriding the write() and create() methods of
all those models, we instead introduced the StudioMixin, and we made
those models use that mixin.