[FIX] quality, quality_mrp: security records
When we are trying to install the quality application, we have a traceback
saying that mrp is not installed.

This problem comes from the security file 'quality.xml' using a mrp
group and quality application doesn't depends on mrp.

To fix this, we moved the security record in the quality_mrp module
which depends on mrp.