[FIX] web_studio: init RPC when entering Studio
An RPC is performed when entering Studio to get the useful information
required by Studio like the dbuuid (needed for the 'Notes' link), a
boolean indicating if there are several translations installed, a
boolean indicating if the chatter can be activated on the edited model.

Some of those information are global (dbuuid, translations) and should
thus be performed only once. This wasn't the case before this rev.
Moreover, those global information might be useful in the app switcher
(in Studio mode), but that RPC was done later on, only when an action
was selected.

This rev. splits the init RPC into two: a global one, which is ensured
to be done only once, the first time entering studio (even from the
app switcher), and an RPC to determine whether or not a chatter can
be activated on the given model, performed each time entering Studio
to edit an action.

This allows to add the 'Notes' link in the app switcher.