[FIX] web: repair length field in view list
Underscore method _.each expects an array like object when a "length"
property is present.

This was an issue with a record having a numeric "length" field set to a
non-negative value. When changing a line the change would not appear on
blur.

This issue was fixed with odoo/odoo@0e664c9e9 but introduced back with
odoo/odoo@f0e331e00.

opw-691070

Cherry-pick of odoo/odoo@5d17749ff47c02294d5ff2ae56bbcef9d082562e