[FIX] account_reports: multi-company in reports
In a previous commit (1c49c6de68d4), a limitation has been introduced in
the account reports. In a multi-company/currency environment, we limit
the company selection to companies which have the same currency than the
user's company currency. The reason was that the amounts were not
correct.

This commit reverts this modification, and fixes the root of the
problem by converting correctly the amounts in the user's company
currency (at the rate of the day).

opw-687512