[FIX] web_studio: properly remove xml editor dirty flag
When a view is changed, the " (unsaved changes)" string is added at the
end of the view name in the <select/> DOM element. As a save of the
view in studio does not reload the whole page, this flag had to be
removed on save.

The implementation to do that only worked when saving a dirty unique
view. If saving a non-dirty view or saving multiple views at a time,
the flag stayed for all view and the name of the selected view
disappeared.

Note: the XML editor is still being refactored for master.