[FIX] web_studio: do not reset changed arch on reselection
In the current implementation, if a view is modified with view editor,
then the user switches to another view and modifies it too, then come
back to the first view... the arch of this view is reset to its origin
and the "unsaved changes" marker is still there.

The code to achieve the right behavior is ugly and incomprehensible
but it has the advantage to keep the intention behind the current
implementation: not adding useless RPC as the website xml editor does.

Note: both xml editor are still being refactored in master.