[FIX] helpdesk: set SLA time fields as required
When searching SLAs and setting order='time_days asc' for example,
it will order them by :
1) lowest to highest time_days that have been set, (e.g.: 0,0,1,2,5,6)
2) all SLAs without time_days set, having null in the database.

What we would like is to have null equivalent to 0 in our search, as the ORM
transforms the null to 0 for Integer fields. Sadly the order in search doesn't
work this way  :(

Thus, time_days, time_hours and time_minutes need to be set and we do this by
making them required fields and adding default=0 for convenience.

Migration
=========

Due to the addition of the required attribute on a field that didn't have one
before, a migration script has been added. It changes all NULLs to 0 for our 3
time fields.