[FIX] web_studio: properly refresh view after multiple xml editions
When editing multiple xml templates thanks to the xml editor and
saving, all the xml templates were properly saved but the view
refresh only condirered a random subset of these changes.

The problem was that the RPCs which save the templates individually
had to be chained so that the last received field_view_get considered
all the changes.
A refresh is still occuring immediately once its related RPC finishes
but subsequent RPCs have to wait for the previous ones to be finished
before starting (and their related refresh will then occur after the
previous ones as binded to the deferred later).