[FIX] helpdesk: set group in view instead of field
The group group_no_one was set in the field declaration instead of on the field
in the view. This created access rights issues when deleting a team for anyone
not in debug mode (except the admin).

It was added to display the field only in debug mode so it's been moved to
the view.