[FIX] web_gantt: gantt tasks order
When we are showing mrp workorders in the gantt view, the tasks order
don't always respect the order in which they have been defined in the
routings.

This is due to the function passed in the gantt 'sort' call, which reorder
the tasks considering their children.

But if both tasks have no children or have children, it returns 0 which
means that the two tasks are equal and so the order between them is not
important. So the order of all elements that are equals depends on the
'Array.prototype.sort()' implementation of the browser.

This causes defferences in the order of tasks in gantt view between Chrome
and Firefox.

To fix this sorting issue, we have to sort the elements having both children
or not, on the original order received from the server.