[FIX] web: don't block if app switcher cancelled
When going to the app switcher, we can be blocked by a modal saying we
have current unsaved changes.

If we go on to the app switcher : everything is fine. But if we don't
confirm the modal warning, we get can get in some odds situations:

- the menus disappears leaving us with a less-than sign,
- we can't click on the app switcher again,
- using escape close then open the warning modal again

This fix (originating from a code by @aab-odoo) should solves the issue.

opw-687435