[FIX] web_studio: don't create ir.model.data twice at import
When records of models customizable in Studio (like ir.model,
ir.model.fields...) are created or edited, we check if the key 'studio'
is in the context, and if so, we create an ir.model.data entry for that
record (if it doesn't exist yet) and flag it with studio.

However, this shouldn't be done when importing a module, as in that
case, an ir.model.data is automatically created afterwards, leading to
duplicate entries in ir.model.data.