[FIX] quality: checks on backorders
An issue arises in the following case:
- Define a quality check on stock location
- Receive a picking to this location (e.g. 5 units)
- Partially receive the picking (e.g. 3 units)
- Perform the quality check
- Transfer the picking, and create a backorder

It is not allowed to transfer the picking since some quality checks are
still pending. These quality checks are created when the stock moves are
split (call to method `split`).

We move the quality check at the very end of the process (end of method
`do_transfer`). At this moment, the backorders have been created, and we
could transfer the remaining quality checks on them.

opw-691906