[FIX] web_studio: properly restore state in form renderer
Before this commit, the set_local_state method in form renderer was a
little too simple.  It tried to restore the currently opened tabs in a
form view, by simply clicking on it (since this is handled by bootstrap)

However, the way it selected each tab to click on was kind of naive:
- the selector was not precise: it could match many anchor tag inside a
  notebook
- it also matched anchor tag without the data-toggle="tab" attribute.
  This is an issue in web studio, since we add a anchor tag with a +
icon to add a tab.  The extra tab was then wrongly clicked on, which
means that we were unable to delete the last page in a notebook.