[FIX] web_studio: add 'studio' key in context of exported data
By setting the 'studio' key to True in the context of the exported
records, they will be correctly flagged 'studio' if the exported module
is then imported on another db, allowing to re-export them from that
other db.

Before this rev., records of an imported module created by Studio
weren't correctly flagged studio.