[FIX] domain of empty/virtual grid cells
read_grid's output is dense (every cell must exist) which means when
read_group has no output for a cell a fake/virtual cell must be
generated, with the relevant information (value) and domain.

The generated domain took in account the relevant grid values for the
row and column, but didn't use the domain with which read_grid itself
was called, possibly leading to way over-selecting cell domains.