[IMP] website_sign: paginate partner selection
The previous way to select partner was adding them as <option/> nodes in
the add "Send" or "Send a copy wizard".

This works well with 100-2000 partners, but over this, a slowness can
make it hard to use or with really too much partners it can break break
the browser.

This commit only does two things:

- using the "ajax" feature of select2 library which allow us to fetch
  data page by page as we want instead of loading all the data in a
  select tag,

- use `<input/>` instead of `<select/>` since in select2 version 3.5 and
  less, the "ajax" feature can only be used with `<input/>`.

The behaviour changes are:

- maximum 30 partners are fetched at a time,

- only 30 partners are displayed at first, but if we scroll to the
  bottom we see the 30 next and so on,

closes #553
opw-691423