[IMP] helpdesk: remove subtype tracking on ticket assignation
Whenever a new user is assigned on a ticket a message is logged in the chatter.

This message was followed by everyone following the ticket by default, which
isn't really desired. Plus the assigned person already gets a custom email
telling him he is assigned to the ticket, so he would get 2 mails for the same
notification.

This commit removes the automatic following of the message.