[FIX] account_reports: allow more unicode date_format
`strftime` of datetime objects returns its result in the current
encoding locale:
 https://docs.python.org/2/library/time.html#time.strftime

Apparently, it also expects it's argument `format` in the current
encoding locale, otherwise an exception can be thrown when trying
to decode it.

In most instance the format is composed of ascii characters, but for
example in chinese date format, it can contain [day]/[month]/[year] in
spelled in chinese which causes an error.

closes #593
opw-694934