[FIX] web_mobile: partner form view: hidden stat button
Before this rev., too many stats buttons were put in the 'More'
dropdown on the partner form view because of the 'Add to contact'
button.

This stat button must be hidden in non-mobile mode. However, as it
is a stat button, it should be hidden using the form view mechanism
(i.e. classname 'o_form_invisible' induced by the 'invisible' property
of the widget) so that the form view can correctly compute the number
of buttons to put in the 'More' dropdown, as it relies on the presence
of the 'o_form_invisible' classname to detect invisible buttons.